Filehash for path (#92)
* exporting resources to the paths based on file hashes

* [dump.to_path]: avoid rewriting existing files

* add default for add-filehash-to-path

* fix lint errors

* fixaps per code review:

* README with more clear explenations and examples
* Avoid copying local file into temp file
* move insert_hash_in_path as method for DumperBase
* insert hash in file path in more relaible way
* avoid code duplication

* typo fix if -> elif

+++ datapackage_pipelines/lib/dump/to_path.py
@@ -9,14 +9,16 @@ class PathDumper(FileDumper):
     def initialize(self, params):
         super(PathDumper, self).initialize(params)
         self.out_path = params.get('out-path', '.')
+        self.add_filehash_to_path = self.__params.get('add-filehash-to-path', False)
         PathDumper.__makedirs(self.out_path)
 
     def write_file_to_output(self, filename, path):
         path = os.path.join(self.out_path, path)
-
         path_part = os.path.dirname(path)
+        # Avoid rewriting existing files
+        if self.add_filehash_to_path and os.path.exists(path_part):
+            return
         PathDumper.__makedirs(path_part)
-
         shutil.copy(filename, path)
         return path
